Extract dependencies module
